feat: move events to separate file
